Resolved Issue #5.  Fully implemented removeEntity() and removed depricateEntity() API methods.
Added a testcase removeEntity()

Added * for the end effector of a template traverse path.  This (*) should be documented as the template traverse path for selecting all nearest neighbors of an entity.
getEntity now raises a Exceptions.NoSuchEntityError(), instead of Exceptions.ScriptError().

Added infrastructure (but not yet a testcase) for the fastsearch option.

NonPersistent Module's Linkrepository getAllLinks() method as appending the UUID of the test Entity to the return link list.  This was not harmful, but bizarre and superflous.  This was removed on the grounds that anything superflous presents a bug risk sometime.