Fixed Issue #22.  Fixed Issue #23.  Fixed Issue #24.  Fixed Issue #25.
Added support for the generic metameme (Graphyne.GenericMetaMeme) to enhance itself.

Fixed a bug in the Graph.SourceMeme.addEnhancement() method, whereby enhancements were only being registered in the source meme and not in the enhancement index; which is where they were actually being used.  This was causing the sourceMemeEnhancementAdd() script method to not work in some cases.

Fixed a bug in itss sister method, the Graph.SourceMeme.removeEnhancement() method, whereby enhancement removals were only being registered in the source meme and not in the enhancement index.

Fixed a bug in EnhancementIndex.removeEnhancement, where enhancement removal was not being registered in the enhancement index.